docs: consolidate examples into root example/ directory
- Replace old config_service.js / config_service_direct.js with config.js
- Add naming.js for service discovery example
- Remove sub-package examples (naming/example, config/example)
- Both examples use the nacos aggregator package, matching real usage
