+++ example/package.json
@@ -1,7 +1,10 @@
 {
   "name": "nacos-example",
   "version": "1.0.0",
+  "private": true,
+  "description": "Usage examples for nacos-sdk-nodejs",
   "dependencies": {
-    "nacos": "^2.5.0"
+    "nacos": "workspace:*",
+    "mz-modules": "^2.1.0"
   }
 }
